complex_rag/requirements.txt
streamlit==1.17.0
torch==1.13.1
datasets==2.8.0
transformers==4.28.0
langchain==0.0.198
sentence-transformers==2.2.2
faiss-cpu==1.7.2
numpy==1.23.5
pandas==1.5.2
scikit-learn==1.1.3
tqdm==4.64.1
re==2.2.1
csv==1.0
